Players can now view Attendance Records. HTML/CSS updated to display pretty on user pages. Raid Loot Fetch function updated for new DAOs. Standard users can now make some AJAX calls. Only 10-20 rows are loaded at a time on a page (instead of 1000s).

+++ api/RaidController.php
@@ -14,7 +14,9 @@ class RaidController {
 			case "PUT":
 				break;
 			case "DELETE":
-				$this->DeleteRow();
+				if(current_user_can("remove_users")) {
+					$this->DeleteRow();
+				}
 				break;
 			default:
 				die();
@@ -54,3 +56,4 @@ class RaidController {
 	private $service;
 }
 add_action('wp_ajax_wro_raid', array(new RaidController(), 'Reroute'));
+add_action('wp_ajax_nopriv_wro_raid', array(new RaidController(), 'Reroute'));
